Redid the project using Visual Studio 2022, and I was able to use HTTPS
locally. For some reason, using just the CLI to generate the project,
I was unable to use HTTPS. HTTP worked fine, but I didn't feel
comfortable moving forward with that. The project was also downgraded to
and empty app since we don't need the full controllers since this is a
REST API.