Added rate limiting and caching.
Calls are limited to 60 per min and 1M per moth.
Before an API call is made, the cache is checked for the data.
If it exists and isn't expired, the cached data is returned.
If it doesn't exist or is expired, the API if allowed and the data
is returned. If We are not allowed to make a call an error message
is returned instead.